+++ src/middlewares/yaml.ts
@@ -15,9 +15,9 @@
 
 import jsYaml, { YAMLException } from "js-yaml";
 import { ParseError } from "../errors/parse";
-import type { HttpMiddleware, HttpRequestHandler, IHttpStringBodyParserOptions, ParseErrorHandler } from "../types";
+import type { HttpRequestHandler, IHttpStringBodyParserOptions, ParseErrorHandler, UniqueHttpMiddleware } from "../types";
 import type { Nilable, Nullable } from "../types/internal";
-import { canHttpMethodHandleBodies, getBufferEncoding, isNil, limitToBytes, readStreamWithLimit, withEntityTooLarge } from "../utils";
+import { canHttpMethodHandleBodies, getBufferEncoding, isNil, limitToBytes, readStreamWithLimit, toUniqueHttpMiddleware, withEntityTooLarge } from "../utils";
 
 interface ICreateMiddlewareOptions {
     encoding: BufferEncoding;
@@ -37,6 +37,11 @@ export interface IYamlOptions extends IHttpStringBodyParserOptions {
 }
 
 /**
+ * Symbol defining the name of this middleware.
+ */
+export const yamlMiddleware: unique symbol = Symbol("yaml");
+
+/**
  * Creates a middleware, that reads the whole input of the request stream,
  * parses it as YAML UTF-8 string and writes the object to 'body' property of the request
  * context.
@@ -46,7 +51,7 @@ export interface IYamlOptions extends IHttpStringBodyParserOptions {
  * @param {Nilable<ParseErrorHandler>} [onParsingFailed] The custom handler, that is invoked, when input is no valid YAML.
  * @param {Nilable<IYamlOptions>} [options] Custom options.
  *
- * @returns {HttpMiddleware} The new middleware.
+ * @returns {UniqueHttpMiddleware} The new middleware.
  *
  * @example
  * ```
@@ -83,10 +88,10 @@ export interface IYamlOptions extends IHttpStringBodyParserOptions {
  * })
  * ```
  */
-export function yaml(): HttpMiddleware;
-export function yaml(limit: number, onLimitReached?: Nilable<HttpRequestHandler>, onParsingFailed?: Nilable<ParseErrorHandler>): HttpMiddleware;
-export function yaml(options: Nilable<IYamlOptions>): HttpMiddleware;
-export function yaml(optionsOrLimit?: Nilable<number | IYamlOptions>, onLimitReached?: Nilable<HttpRequestHandler>, onParsingFailed?: Nilable<ParseErrorHandler>): HttpMiddleware {
+export function yaml(): UniqueHttpMiddleware;
+export function yaml(limit: number, onLimitReached?: Nilable<HttpRequestHandler>, onParsingFailed?: Nilable<ParseErrorHandler>): UniqueHttpMiddleware;
+export function yaml(options: Nilable<IYamlOptions>): UniqueHttpMiddleware;
+export function yaml(optionsOrLimit?: Nilable<number | IYamlOptions>, onLimitReached?: Nilable<HttpRequestHandler>, onParsingFailed?: Nilable<ParseErrorHandler>): UniqueHttpMiddleware {
     if (typeof optionsOrLimit === "number") {
         // [0] number
         // [1] HttpRequestHandler
@@ -136,8 +141,8 @@ export function yaml(optionsOrLimit?: Nilable<number | IYamlOptions>, onLimitRea
     });
 }
 
-function createMiddleware({ encoding, limit, onLimitReached, onParsingFailed }: ICreateMiddlewareOptions): HttpMiddleware {
-    return withEntityTooLarge(async (request, response, next) => {
+function createMiddleware({ encoding, limit, onLimitReached, onParsingFailed }: ICreateMiddlewareOptions): UniqueHttpMiddleware {
+    return toUniqueHttpMiddleware(yamlMiddleware, withEntityTooLarge(async (request, response, next) => {
         try {
             if (canHttpMethodHandleBodies(request.method)) {
                 request.body = jsYaml.loadAll(
@@ -160,5 +165,5 @@ function createMiddleware({ encoding, limit, onLimitReached, onParsingFailed }: 
                 throw error;
             }
         }
-    }, onLimitReached);
+    }, onLimitReached));
 }